Medbot-Project1/requirements.txt
streamlit
langchain-community
openai
nltk
python-dotenv
transformers
torch
emoji
pymongo
pandas
plotly
wordcloud
matplotlib
numpy
